Some bugs fixed for calc and getting values

+++ backend/OLDBRICK-STANJE-ARTIKALA-APP/OLDBRICK-STANJE-ARTIKALA-APP/Entities/Restock.cs
@@ -8,7 +8,7 @@
 
         public int IdPiva { get; set; }
 
-        public decimal Quantity { get; set; }
+        public float Quantity { get; set; }
 
         public DateTime CreatedAt { get; set; }
     }
